separate intersect_plane, intersect_utils
